+++ android/domain/src/main/java/com/domain/edonymyeon/model/Comments.kt
@@ -1,6 +1,6 @@
 package com.domain.edonymyeon.model
 
 data class Comments(
-    val commentCount: Long,
+    val commentCount: Int,
     val comments: List<Comment>,
 )
